refactor(gpu): isolate vendor provider and remove default GPU gesture
